+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @kanso-protocol/accordion
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @kanso-protocol/core@3.0.0
+
 ## 2.0.3
 
 ### Patch Changes