+++ packages/patterns/form-section/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @kanso-protocol/form-section
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @kanso-protocol/core@3.0.0
+  - @kanso-protocol/divider@3.0.0
+
 ## 2.0.3
 
 ### Patch Changes