+++ packages/patterns/nav-item/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @kanso-protocol/nav-item
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @kanso-protocol/core@3.0.0
+  - @kanso-protocol/tooltip@3.0.0
+
 ## 2.0.3
 
 ### Patch Changes