+++ packages/patterns/search-bar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @kanso-protocol/search-bar
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @kanso-protocol/core@3.0.0
+  - @kanso-protocol/button@3.0.0
+  - @kanso-protocol/i18n@3.0.0
+
 ## 2.0.3
 
 ### Patch Changes